매도, 매수 시에 수수료 반영

+++ src/main/scala/com/easywritten/allowancechart/domain/Money.scala
@@ -24,9 +24,9 @@ final case class Money(currency: Currency, amount: MoneyAmount) {
       )
   }
 
-  def *(i: Int): Money = copy(amount = amount * i)
+  def *(i: BigDecimal): Money = copy(amount = amount * i)
 
-  def /(i: Int): Money = copy(amount = amount / i)
+  def /(i: BigDecimal): Money = copy(amount = amount / i)
 
 }
